[dispatch] feat(core): add realtime updates and driver map (wave-1) (#38)
